+++ src/main/java/com/example/gfup2/jwt/UserDetail/UserDetailsServiceImpl.java
@@ -1,7 +1,7 @@
 package com.example.gfup2.jwt.UserDetail;
 
-import com.example.gfup2.domain.member.entity.Member;
-import com.example.gfup2.domain.member.repository.MemberRepository;
+import com.example.gfup2.domain.entity.Member;
+import com.example.gfup2.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
@@ -10,20 +10,14 @@ import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-// userDetailsImple에 account를 넣어주는 서비스입니다.
 public class UserDetailsServiceImpl implements UserDetailsService {
     private final MemberRepository memberRepository;
 
     @Override
     public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
+        Member member = memberRepository.findByemailId(email)
+                .orElseThrow(() -> new UsernameNotFoundException("Not Found Account"));
 
-        Member member = memberRepository.findByemailId(email).orElseThrow(
-                () -> new RuntimeException("Not Found Account")
-        );
-
-        UserDetailsImpl userDetails = new UserDetailsImpl();
-        userDetails.setAccount(member);
-
-        return userDetails;
+        return UserDetailsImpl.from(member);
     }
 }
