+++ src/main/java/com/example/gfup2/jwt/repository/RefreshTokenRepository.java
@@ -3,7 +3,6 @@ package com.example.gfup2.jwt.repository;
 
 import com.example.gfup2.jwt.entity.RefreshToken;
 import org.springframework.data.jpa.repository.JpaRepository;
-
 import java.util.Optional;
 
 public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
