+++ src/main/java/com/example/gfup2/repository/MemberRepository.java
@@ -1,6 +1,6 @@
-package com.example.gfup2.domain.member.repository;
+package com.example.gfup2.repository;
 
-import com.example.gfup2.domain.member.entity.Member;
+import com.example.gfup2.domain.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -8,8 +8,6 @@ import java.util.Optional;
 
 @Repository
 public interface MemberRepository extends JpaRepository<Member, Long> {
-    Optional<Member> findById(Long memberId);
-    Member save(Member member);
     Optional<Member> findByemailId(String email);
 
 }
