+++ src/main/resources/application.properties
@@ -16,8 +16,12 @@ spring.jpa.properties.hibernate.hbm2ddl.auto=update
 logging.level.org.springframework=DEBUG
 
 # AccessTokenPeriod = 1 hour // 1000*60*30 = 1800000
-token.accesstoken-valid-period= 1800000
+token.accesstoken-valid-period= 600
 # RefreshTokenPeriod = 7 days // 1000*60*60*24*7;
 token.refreshtoken-valid-period= 604800000
 
-jwt.secretkey = 7J2064W467Kg7J207IWY7Lqg7ZSE7IiY6rCV7IOd67aE65Ok7ZmU7J207YyF7ZWY7IS47JqU7KKL7J2A7ZqM7IKs7JeQ66qo65GQ7Leo7JeF7ISx6rO17ZWY7Iuk6rGw652866+/7Iq164uI64uk65287J2067iM7IS47IWY65Ok7Ja07KO87IWU7ISc6rCQ7IKs7ZWp64uI64uk64+E7JuA7J2065CY7JeI7Jy866m07KKL6rKg7Iq164uI64uk
+jwt.secretkey = 7J2064W467Kg7J207IWY7Lqg7ZSE7IiY6rCV7IOd67aE65Ok7ZmU7J207YyF7ZWY7IS47JqU7K
+
+# logging
+logging.file.name=./logs/app.log
+logging.level.root=DEBUG
